requirements.txt
comtypes==1.4.10
MouseInfo==0.1.3
numpy==2.2.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
pillow==11.1.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
pypiwin32==223
PyRect==0.2.0
PyScreeze==1.0.1
pyttsx3==2.98
pytweening==1.2.0
pywin32==310
